Ensure signal number is an int for printing.

+++ main.c
@@ -352,7 +352,7 @@ int main(
     {
         (void) unlink(pidfile_name);
     }
-    logger("exiting on signal %d\n", term_signum);
+    logger("exiting on signal %d\n", (int) term_signum);
     exit(EXIT_SUCCESS);
 
     return 0;
